fix protected blogs page, now is using client caching with tanstack!
