+++ src/app/(protected-members-routes)/_components/nav-components/available-links.tsx
@@ -24,11 +24,11 @@ export const AVAILABLE_LINKS = [
                 href: '/members/blogs',
                 icon: <ScrollText className="shrink-0" />,
             },
-            {
-                title: 'Blogss',
-                href: '/members/blogss',
-                icon: <ScrollText className="shrink-0" />,
-            },
+            // {
+            //     title: 'Blogss',
+            //     href: '/members/blogss',
+            //     icon: <ScrollText className="shrink-0" />,
+            // },
             // {
             //     title: 'Practice Areas',
             //     href: '/members/practice-areas',
